+++ graph/src/main/java/net/spals/appbuilder/graph/writer/NoOpServiceGraphWriterPlugin.java
@@ -1,14 +1,15 @@
 package net.spals.appbuilder.graph.writer;
 
-import net.spals.appbuilder.graph.model.ServiceDAG;
+import net.spals.appbuilder.graph.model.IServiceGraphVertex;
+import net.spals.appbuilder.graph.model.ServiceGraph;
 
 /**
  * @author tkral
  */
-class NoOpServiceGraphWriterPlugin implements ServiceGraphWriterPlugin {
+class NoOpServiceGraphWriterPlugin implements ServiceGraphWriterPlugin<IServiceGraphVertex<?>, ServiceGraph> {
 
     @Override
-    public String writeServiceGraph(final ServiceDAG serviceDAG) {
+    public String writeServiceGraph(final ServiceGraph serviceGraph) {
         return "Skipping service graph write...";
     }
 }
