+++ graph/src/main/java/net/spals/appbuilder/graph/writer/ServiceGraphWriterPlugin.java
@@ -1,11 +1,13 @@
 package net.spals.appbuilder.graph.writer;
 
-import net.spals.appbuilder.graph.model.ServiceDAG;
+import net.spals.appbuilder.graph.model.IServiceGraphVertex;
+import org.jgrapht.DirectedGraph;
+import org.jgrapht.graph.DefaultEdge;
 
 /**
  * @author tkral
  */
-interface ServiceGraphWriterPlugin {
+interface ServiceGraphWriterPlugin<V extends IServiceGraphVertex<?>, G extends DirectedGraph<V, DefaultEdge>> {
 
-    String writeServiceGraph(ServiceDAG serviceDAG);
+    String writeServiceGraph(G serviceGraph);
 }
